Refactor: AddPlan 폴더 생성
